fix[]: fix up the datatypes to fit the task description

+++ oving10/src/ooving10_oppg1/Arrangement.java
@@ -1,16 +1,16 @@
 package ooving10_oppg1;
 
 public class Arrangement {
-    private final long nummer;
+    private final int nummer;
     private final String navn;
     private final String sted;
     private final String arrangoer;
     private final String type;
-    private final int tidspunkt;
+    private final long tidspunkt;
 
 
-    public Arrangement(long nummer, String navn, String sted, 
-                       String arrangoer, String type, int tidspunkt){
+    public Arrangement(int nummer, String navn, String sted, 
+                       String arrangoer, String type, long tidspunkt){
         this.nummer = nummer;
         this.navn = navn;
         this.sted = sted;
@@ -27,12 +27,12 @@ public class Arrangement {
     );
 }
 
-    public long getNummer(){ return nummer;}
+    public int getNummer(){ return nummer;}
     public String getNavn(){ return navn;}
     public String getSted(){ return sted;}
     public String getArrangoer(){ return arrangoer;}
     public String getType(){ return type;}
-    public int getTidspunkt(){ return tidspunkt;}
+    public long getTidspunkt(){ return tidspunkt;}
